FEAT: Add windows to the engine
This commit adds the window class which should be fully platform
independent. The class holds an ID based on the previously added
ID-type class. This ID represents an actual instance of a platform
dependent window. Currently a platform dependent window has been
implemented with the use of thirdparty software library GLFW, which
should be supported on multiple platforms, without changes to the code.

No known issues, however GLFW also requires event handling to be
implemented