FEAT: Add events and callback to window
This commit takes the previously added event-system and adds it to the
current window-system (GLFW). Several GLFW-callbacks are now setup to
link with the new event-system
    - Added event-system callbacks for window close, window resize etc
    - Added new GLFW callbacks for key-events and mouse-events and
      linked them to the event-system
    - Changed GLFW close-window callback to instead hide the window and
      set internal variable isClosed to true
    - Changed the GLFW createWindow to have an context (OpenGL)
      attached to it, such that each window can be updated each frame

Minor changes to the already existing window functionality
    - Updated naming of variables to better match other place in the
      engine
    - Added todo comments for future adding of a freelist-container
      to be used instead of the vector and deque currently inplace

Update the tests in the muggyExample build such that we are now
utilizing the event-system

When testing the implementation there seem to be some sort of assertion
happening when the program ends. There is currently no fix for this.