FEAT: Add game_entity and transform
This commit marks the start of the entity system to be used in the
engine. This defines the game_entity class which describes any entity
in the game (camera, NPC, tree, player etc). The actual class is
supposed to be open to the user and is therefor defined inside
engineAPI-folder
    - Adds the game_entity implementation based on the previously added
      id-implementation from id.h
    - Adds the implementation of entity_id which takes care of a valid
      or non-valid entity

This commit also adds the first component to game_entity which is its
transform, describing the entitys position, and is a requirement for
all entities to have.

No known side effects