TEST: Tests functionality of entities
This commit adds testing of the game_entities system added previously
    - Added new header- and cpp-files for the actual test
    - Added new test-defines for the muggyExample project

This test will bulk-add 1000 entities to the game_entity system at the
start of the test and then add/remove a random number of entities for
each individual iteration. Every 10 000 iteration, the current
statistics (added, remove, total) are printed to the console

With the current id-setup and the generation set to 8-bits, the system
will roll-over after sometime (expect ~700 000 added/removed entities),
causing assertions from the id-header to hit. This is expected and part
of the current design