FEAT: Add free_list container
This commit adds the free_list container class which at the core uses
a vector as data container and then adds additional functionality ontop
such that the free_list class acts as an array with holes
    - The free_list class keeps track of "free" slots within itself
      such that memory can be reused
    - This class currently uses the std::vector implementation

TODO: Add own vector implementation to avoid calling the items
destructor twice

The current implementation, which uses std::vector, results in two
calls to an items destructor. One, when the free_list removes an item
and one when the free_list itself is removed. This second call
currently results in an assertion when the application ends