TEST: Tests the free_list functionality
This commit is the first part of adding testing of the free_list
container, to verify its functionality. Please note that the tests are
not complete
    - Added new header- and cpp-fiels for the actual test
    - Added new test-defines for the muggyExample project

This commit also contains minor changes to the window tests of the
muggyExample project, and small changes to types in the underlying
platform layer for the glfw-to-event implementation.
This commit also contains minor changes to the vector class container,
specifically removal of the iterator implementation, which does not
work