FEAT: Add script-support to game_entity
This commit adds the script-component to the entity-system. Currently
the scripts supported are only native C++ based scripts. A script can
be used to describe how the connected entity should behave, for example
a player
    - Adds a script as a component of an entity
    - An entity can only have one script component but does not need
      to have a script

This commit also adds minor changes to entity and the transform-
component to support the new script-component.
This commit also adds a math-conversion function for euler angles to
quaternions, in order to support different inputs to the engine