CHORE: Change buildsystem to premake5
This project have used cmake so far as a buildsystem. However, I am not
enjoying how the generation works in windows and specifically when
using it to generate for Visual Studio solutions. Because of this I am
opting to try out premake instead of cmake
    - Premake is a little bit easier to setup than cmake
    - Premake allows for generation of the solution within the build
      directory, which might ofcourse be a bad thing, however currently
      this is just a benefit
    - The premake setup still uses cmake to build the thirdparty glfw
      library, however this is as simple as invoking cmake on the
      already available cmake setup from glfw

As a side note, this is the first commit in the dev-branch which will
be the "new" main-branch for development. When dev-branch is stable it
will then be merged with main to allow main to stay relatively clean
and only contain working releases